Fix intermittent test failure in test_deny_executable_write (#1588)
The test was writing 42 to offset 4 of the program account. The program
is upgradable, so offset 4 is first byte of the account which holds the
actual elf program.

There was a chance that the first byte of the account is 42, in which
case the test passes, because with the non-direct mapping case, you can
write to read only accounts as long as the contents does not change.