cargo-build-sbf: Use `metadata.solana.tools-version` in Cargo.toml (#2914)
* cargo-build-sbf: Use `solana.tools-version` in Cargo.toml

#### Problem

It's possible to specify the version of the platform-tools version using
the `--tools-version`, but most people aren't aware of the different
versions of the compiler, or that they can use the flag.

#### Summary of changes

Allow `cargo-build-sbf` to read from the metadata section of a package
or workspace Cargo.toml. The concept is similar to
`rust-toolchain.toml`, where cargo finds the right version of Rust. For
cargo-build-sbf, it can read:

```toml
[package.metadata.solana]
tools-version = "v1.43"
```

Or

```toml
[workspace.metadata.solana]
tools-version = "v1.43"
```

To go with this change, since I often forget leading "v" in the version
string, the version parsing now allows for omitting the leading "v", so
you can specify `--tools-version 1.43`.

* Relax leading "v"

* Add CHANGELOG entry

* Special-case default run

* Add tests

* ci: Fix parsing platform tools version