PrioritizationFeeCache: make update() never sleep on the sender channel (#3813)
Avoid update() callers having to notify the service_loop() thread by
doing an explicit sleep instead of sleeping on channel.recv() when the
channel is empty. This avoids the case in which multiple replay/banking
threads call update() at the same time and end up... sleeping themselves
acquiring the mutex to notify the receiver.